requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg==3.1.18
python-dotenv==1.0.0